final fix

+++ sivkov.filipp/P4/localMax.cpp
@@ -2,7 +2,7 @@
 
 int localMaxCounter(int* m, size_t rows, size_t cols, int count, int numOfTask)
 {
-  for (int i = 0; i < cols * rows - 1; i++)
+  for (size_t i = 0; i < cols * rows - 1; i++)
   {
     if ((i > cols) && (i < rows * cols - cols) && ((i % cols) != 0 && (i % rows) != 0))
     {
